KiroVibe/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
chromadb==0.4.24
sentence-transformers==2.7.0
huggingface-hub==0.23.0
python-dotenv==1.0.1
boto3==1.35.0
langchain-aws==0.1.0
opensearch-py==2.4.0
faiss-cpu==1.8.0